Create page release cron script

+++ src/ATPComic/Model/Page.php
@@ -13,49 +13,54 @@ class Page extends \ATP\ActiveRecord
 	
 	public function postSave()
 	{
-		$newArcIds = $this->arcAssignments;
-		$nodes = (array)$this->getNodes();
-		
-		//Determine nodes to delete
-		$nodesToDelete = array_filter($nodes, function($node) use ($newArcIds) {
-			return !in_array($node->arcId, $newArcIds);
-		});
-		
-		//Determine arcs to add
-		$arcsToAdd = array_filter($newArcIds, function($id) use ($nodes) {
-			foreach($nodes as $node)
-			{
-				if($node->arcId == $id) return false;
-			}
-			return true;
-		});
-		
-		//Delete unneeded nodes
-		foreach($nodesToDelete as $node)
+		if(isset($this->arcAssignments))
 		{
-			//Delete the node
-			$node->delete();
+		
+			$newArcIds = $this->arcAssignments;
+			$nodes = (array)$this->getNodes();
 			
-			//Update page numbers on subsequent pages
-			$db = $node->getAdapter();
-			$sql = "UPDATE atpcomic_nodes SET page_number = page_number-1 WHERE page_number > ? AND arc_id = ?";
-			$db->query($sql, array($node->pageNumber, $node->arcId));
-		}
-		
-		//Add new arcs
-		foreach($arcsToAdd as $arcId)
-		{			
-			//Get the last node in the arc
-			$arc = new \ATPComic\Model\Arc();
-			$arc->loadById($arcId);
-			$lastNode = $arc->lastNode();
-		
-			$newNode = new \ATPComic\Model\Node();
-			$newNode->arc = $arc;
-			$newNode->page = $this;
-			$newNode->pageNumber = $lastNode ? $lastNode->pageNumber + 1 : 1;
-		
-			$newNode->save();
+			//Determine nodes to delete
+			$nodesToDelete = array_filter($nodes, function($node) use ($newArcIds) {
+				return !in_array($node->arcId, $newArcIds);
+			});
+			
+			//Determine arcs to add
+			$arcsToAdd = array_filter($newArcIds, function($id) use ($nodes) {
+				foreach($nodes as $node)
+				{
+					if($node->arcId == $id) return false;
+				}
+				return true;
+			});
+			
+			//Delete unneeded nodes
+			foreach($nodesToDelete as $node)
+			{
+				//Delete the node
+				$node->delete();
+				
+				//Update page numbers on subsequent pages
+				$db = $node->getAdapter();
+				$sql = "UPDATE atpcomic_nodes SET page_number = page_number-1 WHERE page_number > ? AND arc_id = ?";
+				$db->query($sql, array($node->pageNumber, $node->arcId));
+			}
+			
+			//Add new arcs
+			foreach($arcsToAdd as $arcId)
+			{			
+				//Get the last node in the arc
+				$arc = new \ATPComic\Model\Arc();
+				$arc->loadById($arcId);
+				$lastNode = $arc->lastNode();
+			
+				$newNode = new \ATPComic\Model\Node();
+				$newNode->arc = $arc;
+				$newNode->page = $this;
+				$newNode->pageNumber = $lastNode ? $lastNode->pageNumber + 1 : 1;
+				$newNode->isActive = $this->isActive;
+			
+				$newNode->save();
+			}
 		}
 		
 		//Update active flag on nodes
